fix(monitor): cast ::float8 nas queries de health_checks (numeric → f64 panic)
Bug em produção: aba VM > section "Health checks externos" travava em
loading infinito. Investigado via RUST_LOG=info revelou panic:

  thread 'tokio-rt-worker' panicked at queries.rs:294:25:
  error retrieving column 0: error deserializing column 0

Causa raíz: avg(avg_response_ms) e (100.0 * SUM/NULLIF) retornam tipo
"numeric" no Postgres, e tokio_postgres não converte numeric → f64
sem o decimal crate. Fix: cast explícito ::float8 nas 4 queries
(uptime 24h/7d/30d + avg response 24h).

Bonus: o panic engolido pelo Tauri runtime virava promise pendente
no JS — usePolling nunca recebia data nem error → loading infinito.
A versão anterior do PR #7 já adicionou error surfacing no component,
mas o panic não vira reject por dentro do Tauri (só nunca resolve).
Agora que o panic some, o command resolve normal com dados.

+++ src-tauri/src/monitor/queries.rs
@@ -251,7 +251,7 @@ pub async fn fetch_health_summary(pool: &Pool, endpoint: &str) -> Result<HealthC
     // Uptime 24h vem do continuous aggregate horário
     let uptime_24h = client
         .query_opt(
-            "SELECT 100.0 * SUM(up)::float / NULLIF(SUM(total), 0) AS pct
+            "SELECT (100.0 * SUM(up)::float / NULLIF(SUM(total), 0))::float8 AS pct
              FROM health_checks_hourly
              WHERE endpoint = $1 AND bucket > NOW() - INTERVAL '24 hours'",
             &[&endpoint],
@@ -263,7 +263,7 @@ pub async fn fetch_health_summary(pool: &Pool, endpoint: &str) -> Result<HealthC
     // Uptime 7d / 30d vem do agregado diário
     let uptime_7d = client
         .query_opt(
-            "SELECT 100.0 * SUM(up)::float / NULLIF(SUM(total), 0) AS pct
+            "SELECT (100.0 * SUM(up)::float / NULLIF(SUM(total), 0))::float8 AS pct
              FROM health_checks_daily
              WHERE endpoint = $1 AND bucket > NOW() - INTERVAL '7 days'",
             &[&endpoint],
@@ -274,7 +274,7 @@ pub async fn fetch_health_summary(pool: &Pool, endpoint: &str) -> Result<HealthC
 
     let uptime_30d = client
         .query_opt(
-            "SELECT 100.0 * SUM(up)::float / NULLIF(SUM(total), 0) AS pct
+            "SELECT (100.0 * SUM(up)::float / NULLIF(SUM(total), 0))::float8 AS pct
              FROM health_checks_daily
              WHERE endpoint = $1 AND bucket > NOW() - INTERVAL '30 days'",
             &[&endpoint],
@@ -285,7 +285,10 @@ pub async fn fetch_health_summary(pool: &Pool, endpoint: &str) -> Result<HealthC
 
     let avg_response_ms_24h = client
         .query_opt(
-            "SELECT avg(avg_response_ms) FROM health_checks_hourly
+            // ::float8 explicit cast — avg(numeric) volta numeric, e tokio_postgres
+            // não converte numeric → f64 sem o decimal crate. Sem o cast, panicou
+            // em "error deserializing column 0".
+            "SELECT avg(avg_response_ms)::float8 FROM health_checks_hourly
              WHERE endpoint = $1 AND bucket > NOW() - INTERVAL '24 hours'",
             &[&endpoint],
         )
